fix: implement Willie's suggestions
